Fix 59 Dart compilation errors for Phases 6-7 screens
- Added missing route constants in app_routes.dart
- Fixed getter syntax and type conversions in decentralized_storage_screen.dart
- Fixed const extension method errors in federated_ai_screen.dart
- Fixed type mismatches and undefined identifiers in fer_chain_screen.dart
- Fixed undefined variables in ferret_calendar_screen.dart
- Replaced undefined Colors constants in ferret_mail_screen.dart
- Fixed logical operators in ferret_terminal_screen.dart
- Added 'late' keyword to uninitialized game class fields (9 game files)
- Renamed Theme classes for Flutter 3.41.9 compatibility in app_theme.dart
- Added dart:math import and fixed type casting in aifer_hud.dart
- Verified: flutter analyze shows 0 errors

+++ lib/widgets/aifer_hud.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:sizer/sizer.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'dart:async';
+import 'dart:math';
 
 /// AiFERHUD - Heads-Up Display overlay for system status
 /// This is a placeholder widget for future implementation
@@ -99,10 +100,10 @@ class _AiFERHUDState extends State<AiFERHUD>
       if (mounted) {
         setState(() {
           _batteryLevel = (_batteryLevel + (Random().nextInt(3) - 1))
-              .clamp(0, 100);
+              .clamp(0, 100).toInt();
           _networkStrength = (_networkStrength + (Random().nextInt(5) - 2))
-              .clamp(0, 100);
-          _meshNodes = (_meshNodes + (Random().nextInt(3) - 1)).clamp(0, 100);
+              .clamp(0, 100).toInt();
+          _meshNodes = (_meshNodes + (Random().nextInt(3) - 1)).clamp(0, 100).toInt();
         });
       }
     });
